Layout instead of page in all app pages
