+++ pages/admin/pages/new.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import PageForm from 'components/admin/pages/form/PageForm';
 import Title from 'components/ui/Title';
-import Page from 'components/admin/layout/Page';
+import Layout from 'components/admin/layout/Layout';
 
 export default class PageNew extends React.Component {
 
   render() {
     return (
-      <Page
+      <Layout
         title="New page"
         description="Edit page description..."
       >
@@ -16,7 +16,7 @@ export default class PageNew extends React.Component {
             <Title className="-huge -p-primary">
               Create Static Page
             </Title>
-            <PageForm
+            <LayoutForm
               application={['rw']}
               authorization={process.env.TEMP_TOKEN}
               onSubmit={() => Router.pushRoute('pages')}
@@ -24,7 +24,7 @@ export default class PageNew extends React.Component {
             />
           </div>
         </div>
-      </Page>
+      </Layout>
     );
   }
 }
