+++ pages/app/GetInvolved.js
@@ -7,6 +7,7 @@ import Banner from 'components/app/common/Banner';
 import Intro from 'components/app/common/Intro';
 import CardStatic from 'components/app/common/CardStatic';
 import Page from 'components/app/layout/Page';
+import Layout from 'components/app/layout/Layout';
 
 const cards = [
   {
@@ -68,7 +69,7 @@ const cards = [
   }
 ];
 
-class GetInvolved extends React.Component {
+class GetInvolved extends Page {
   componentWillMount() {
     this.props.getStaticData('get-involved', 'getInvolved');
   }
@@ -106,7 +107,7 @@ class GetInvolved extends React.Component {
     );
 
     return (
-      <Page
+      <Layout
         title="Get Involved"
         description="Get Involved description"
       >
@@ -135,7 +136,7 @@ class GetInvolved extends React.Component {
             </div>
           </div>
         </div>
-      </Page>
+      </Layout>
     );
   }
 }
