+++ pages/app/Home.js
@@ -5,6 +5,7 @@ import CardStatic from 'components/app/common/CardStatic';
 import Rating from 'components/app/common/Rating';
 import Icon from 'components/ui/Icon';
 import Page from 'components/app/layout/Page';
+import Layout from 'components/app/layout/Layout';
 import MoveTo from 'moveto';
 
 const insightsCards = [
@@ -73,7 +74,7 @@ const exploreCards = [
   }
 ];
 
-class Home extends React.Component {
+class Home extends Page {
   componentDidMount() {
     this.setAnchorScroll('discoverIsights', 'js-scroll');
   }
@@ -135,9 +136,10 @@ class Home extends React.Component {
     const insightsCardsStatic = this.insightsCardsStatic();
 
     return (
-      <Page
+      <Layout
         title="Resource Watch"
         description="Resource Watch description"
+        user={this.props.user}
       >
         <div className="p-home">
           <div className="c-page">
@@ -248,7 +250,7 @@ class Home extends React.Component {
             </Banner>
           </div>
         </div>
-      </Page>
+      </Layout>
     );
   }
 }
