+++ pages/app/Partners.js
@@ -7,13 +7,14 @@ import { getPartners } from 'redactions/partners';
 
 // Components
 import Page from 'components/app/layout/Page';
+import Layout from 'components/app/layout/Layout';
 import Title from 'components/ui/Title';
 import PartnerBlock from 'components/app/common/Partners/PartnerBlock';
 import Banner from 'components/app/common/Banner';
 import Breadcrumbs from 'components/ui/Breadcrumbs';
 
 
-class Partners extends React.Component {
+class Partners extends Page {
   componentWillMount() {
     this.props.getPartners();
   }
@@ -23,7 +24,7 @@ class Partners extends React.Component {
     const funders = this.props.list.filter(p => p.attributes['partner-type'] === 'funders');
 
     return (
-      <Page
+      <Layout
         title="Partners"
         description="Partners description"
       >
@@ -117,7 +118,7 @@ class Partners extends React.Component {
             </div>
           </div>
         </div>
-      </Page>
+      </Layout>
     );
   }
 }
