+++ pages/app/Pulse.js
@@ -1,6 +1,14 @@
 import React from 'react';
 import { Autobind } from 'es-decorators';
 
+// Redux
+import withRedux from 'next-redux-wrapper';
+import { initStore } from 'store';
+import { getLayers, getLayerPoints } from 'redactions/pulse';
+import getLayersGroupPulse from 'selectors/pulse/layersGroupPulse';
+import getActiveLayersPulse from 'selectors/pulse/layersActivePulse';
+import { toggleTooltip } from 'redactions/tooltip';
+
 // Helpers
 import LayerGlobeManager from 'utils/layers/LayerGlobeManager';
 
@@ -13,18 +21,13 @@ import Spinner from 'components/ui/Spinner';
 import ZoomControl from 'components/ui/ZoomControl';
 import GlobeTooltip from 'components/app/pulse/GlobeTooltip';
 import Page from 'components/app/layout/Page';
-import withRedux from 'next-redux-wrapper';
-import { initStore } from 'store';
-import { getLayers, getLayerPoints } from 'redactions/pulse';
-import getLayersGroupPulse from 'selectors/pulse/layersGroupPulse';
-import getActiveLayersPulse from 'selectors/pulse/layersActivePulse';
-import { toggleTooltip } from 'redactions/tooltip';
+import Layout from 'components/app/layout/Layout';
 
 const earthImage = 'static/images/components/vis/earth-min.jpg';
 const earthBumpImage = 'static/images/components/vis/earth-bump.jpg';
 const cloudsImage = 'static/images/components/vis/clouds-min.png';
 
-class Pulse extends React.Component {
+class Pulse extends Page {
 
   static async getInitialProps({ pathname }) {
     return { pathname };
@@ -219,7 +222,7 @@ class Pulse extends React.Component {
     const globeWidht = (typeof window === 'undefined') ? 500 : window.innerWidth;
     const globeHeight = (typeof window === 'undefined') ? 300 : window.innerHeight - 130; // TODO: 130 is the header height
     return (
-      <Page
+      <Layout
         title="Planet Pulse"
         description="Planet Pulse description"
         pathname={pathname}
@@ -265,7 +268,7 @@ class Pulse extends React.Component {
             onZoomOut={this.triggerZoomOut}
           />
         </div>
-      </Page>
+      </Layout>
     );
   }
 }
